metrics: density: add memory footprint test
Add a test that measures the average PSS across a number of running
containers. The test takes into account all the components of the
VM container system (VM, proxy and shim), and spreads the cost of
the shared proxy executable across all running containers.
The test can also insert a delay between launching the containers and
taking the PSS measurement to allow any system memory optimisations
to settle, such as from KSM

Fixes: #383

Signed-off-by: Graham Whaley <graham.whaley@intel.com>