metrics: iperf3: pps: add ability to iterate the test
Quite often we want to run a test multiple times to get a
stable average result. To do this efficiently, restructure the
iperf3 cnt cnt pps test to launch and then re-use two containers
rather than launch and kill them for each iteration.

This is a win on speed, and tests have shown that the variablility
of the results is the same.

Signed-off-by: Graham whaley <graham.whaley@intel.com>